feat: added fast filters
